fix NSDictionary+Merge

+++ Categories/UIKit/UIImage/UIImage+FX.m
@@ -301,7 +301,9 @@
 - (UIImage *)imageWithShadowColor:(UIColor *)color offset:(CGSize)offset blur:(CGFloat)blur
 {
     //get size
-    CGSize border = CGSizeMake(fabsf(offset.width) + blur, fabsf(offset.height) + blur);
+    //CGSize border = CGSizeMake(fabsf(offset.width) + blur, fabsf(offset.height) + blur);
+    CGSize border = CGSizeMake(fabs(offset.width) + blur, fabs(offset.height) + blur);
+
     CGSize size = CGSizeMake(self.size.width + border.width * 2.0f, self.size.height + border.height * 2.0f);
     
     //create drawing context
